Add times config
Each time (level) has a simple config file, which defines the
background, obstacles, weather, etc. times/config.js defines the
transitions between the times.

The times config is meant to be used in the time machine while traveling back and forth.
The config of each time can be used in the main state to setup objects
(Background, Obstacle, etc.).